+++ ripozo/wsgi/__init__.py
@@ -1,3 +1,6 @@
+"""
+Utilities for parsing a WSGI environ
+"""
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function